Layout

+++ app/controllers/tutor_controller.rb
@@ -148,7 +148,11 @@ class TutorController < ApplicationController
   def toggle_is_live
     if current_user.is_live
       current_user.update_attributes(is_live: false)
-      current_user.location.destroy
+
+      unless current_user.location.nil?
+        current_user.location.destroy
+      end
+      
       respond_to do |format|
         format.js { render 'offline'}
       end
